requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.0
cfgv==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distlib==0.4.0
Django==4.2.27
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
filelock==3.19.1
flake8==7.3.0
identify==2.6.15
importlib_metadata==8.7.0
isort==6.1.0
kombu==5.6.1
mccabe==0.7.0
nodeenv==1.9.1
packaging==25.0
platformdirs==4.4.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
virtualenv==20.35.4
wcwidth==0.2.14
zipp==3.23.0
